+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Configurable OPA executable path through optional `OPA_PATH` environment variable ([#1](https://github.com/johanfylling/opa-dependency-manager/issues/1))
 - Refactored `opa.project` yaml structure ([#4](https://github.com/johanfylling/opa-dependency-manager/pull/14))
 - Info/debug printing to `stderr` ([#6](https://github.com/johanfylling/opa-dependency-manager/issues/6))
+- Added `build` command for building OPA bundles ([#8](https://github.com/johanfylling/opa-dependency-manager/issues/8))
 
 ## [0.1.0]